refactor(app): Changed the models for ResponsiveImage and RetinaImage to allow for more flexible ima
Previously the keys for the ResponsiveImage object were the string representations of the breakpoint
sizes. This meant that any GraphQL server providing that image object would need to be aware of
those keys in order to provide an appropriate object. With the strings abstracted into a sizes
property, and retina images seperated from palceholder and dummy, the responsiveImage object is more
flexible.

BREAKING CHANGE: sn-image-loader component now expects a differently structured responsiveImage
object.